Replace Koin with Hilt.
