+++ app/build.gradle
@@ -6,7 +6,9 @@ apply plugin: 'kotlin-android-extensions'
 
 apply plugin: 'androidx.navigation.safeargs.kotlin'
 
+apply plugin: 'dagger.hilt.android.plugin'
 apply plugin: 'kotlin-kapt' // For Kotlin, explicitly apply the kapt plugin (apply plugin: 'kotlin-kapt') before the ObjectBox plugin.
+
 apply plugin: 'io.objectbox' // Apply last.
 
 android {
@@ -25,6 +27,10 @@ android {
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
     }
 
     buildTypes {
@@ -64,8 +70,8 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.2.2'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.2.2'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.0'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.0'
     // Net
     implementation 'com.google.code.gson:gson:2.8.6'
     implementation 'com.squareup.retrofit2:retrofit:2.7.2'
@@ -81,10 +87,12 @@ dependencies {
     // Timber
     implementation "com.jakewharton.timber:timber:4.7.1"
 
-    // Koin
-    implementation "org.koin:koin-android:2.1.2"
-    implementation "org.koin:koin-androidx-viewmodel:2.1.2"
-    implementation "org.koin:koin-core:2.1.2"
+    // Hilt dagger
+    implementation "com.google.dagger:hilt-android:2.28-alpha"
+    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
+    implementation 'androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha01'
+    kapt 'androidx.hilt:hilt-compiler:1.0.0-alpha01'
+
 
     // ObjectBox
     implementation "io.objectbox:objectbox-kotlin:2.5.1"
@@ -105,6 +113,5 @@ dependencies {
     testImplementation 'junit:junit:4.13'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    testImplementation 'org.koin:koin-test:2.1.2'
     testImplementation 'org.mockito:mockito-inline:2.21.0'
 }
