+++ app/src/main/java/com/example/simpleandroidapp/App.kt
@@ -3,10 +3,6 @@ package com.example.simpleandroidapp
 import android.annotation.SuppressLint
 import android.app.Application
 import android.util.Log
-import com.example.simpleandroidapp.net.networkModule
-import com.example.simpleandroidapp.repository.beerModule
-import com.example.simpleandroidapp.repository.dao.boxModule
-import com.example.simpleandroidapp.ui.second.beerViewModule
 import com.facebook.flipper.android.AndroidFlipperClient
 import com.facebook.flipper.android.utils.FlipperUtils
 import com.facebook.flipper.core.FlipperClient
@@ -14,24 +10,13 @@ import com.facebook.flipper.plugins.inspector.DescriptorMapping
 import com.facebook.flipper.plugins.inspector.InspectorFlipperPlugin
 import com.facebook.flipper.plugins.network.NetworkFlipperPlugin
 import com.facebook.soloader.SoLoader
-import org.koin.android.ext.koin.androidContext
-import org.koin.android.ext.koin.androidLogger
-import org.koin.core.context.startKoin
+import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
 
+@HiltAndroidApp
 class App : Application() {
     override fun onCreate() {
         super.onCreate()
-
-        startKoin {
-            androidLogger()
-            androidContext(this@App)
-            modules(listOf(
-                networkModule,
-                boxModule,
-                beerModule,
-                beerViewModule))
-        }
 
         SoLoader.init(this, false)
