+++ app/src/main/java/com/example/simpleandroidapp/repository/dao/ObjectBox.kt
@@ -1,22 +1,26 @@
 package com.example.simpleandroidapp.repository.dao
 
 import android.app.Application
+import android.content.Context
 import com.example.simpleandroidapp.BuildConfig
 import com.example.simpleandroidapp.repository.pojo.MyObjectBox
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+import dagger.hilt.android.qualifiers.ApplicationContext
 import io.objectbox.BoxStore
 import io.objectbox.android.AndroidObjectBrowser
-import org.koin.dsl.module
 import timber.log.Timber
+import javax.inject.Singleton
 
-val boxModule = module {
-    single { ObjectBox.init(get()) }
-}
+
 
 object ObjectBox {
     lateinit var boxStore: BoxStore
         private set
 
-    fun init(applicationContext: Application): BoxStore {
+    fun init(applicationContext: Context): BoxStore {
         if (::boxStore.isInitialized && !boxStore.isClosed) {
             return boxStore
         }
